backend

+++ Week-6/express-server/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 const { text } = require('express');
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler ( async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({user: req.user.id});
 
     res.status(200).json(goals);
 })
@@ -22,6 +23,7 @@ const setGoal = asyncHandler ( async (req, res) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     })
     
@@ -37,6 +39,20 @@ const updateGoal = asyncHandler ( async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error("Goal not found")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check if the user exists
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // Check if the user is the owner of the goal
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +71,20 @@ const deleteGoal = asyncHandler ( async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error("Goal not found")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check if the user exists
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // Check if the user is the owner of the goal
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     await goal.deleteOne();
